LayoutTests/fast/css/parsing-css-attribute-case-insensitive-value-4-expected.txt

This test verify that invalid case-insenstivie attribute value matching (e.g. [foo=bar i]) do not affect surrounding rules.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL getComputedStyle(document.getElementById('target')).color should be rgb(0, 255, 0). Was rgb(0, 0, 0).
FAIL getComputedStyle(document.getElementById('target')).backgroundColor should be rgb(255, 0, 0). Was rgba(0, 0, 0, 0).
FAIL document.getElementById('style-container').sheet.cssRules.length should be 2. Was 0.
FAIL document.getElementById('style-container').sheet.cssRules[0].selectorText should be :matches(#notthere >> div, #target). Threw exception TypeError: undefined is not an object (evaluating 'document.getElementById('style-container').sheet.cssRules[0].selectorText')
FAIL document.getElementById('style-container').sheet.cssRules[0].cssText should be :matches(#notthere >> div, #target) { background-color: red; }. Threw exception TypeError: undefined is not an object (evaluating 'document.getElementById('style-container').sheet.cssRules[0].cssText')
FAIL document.getElementById('style-container').sheet.cssRules[1].selectorText should be :matches(#notthere >> foobar, #target). Threw exception TypeError: undefined is not an object (evaluating 'document.getElementById('style-container').sheet.cssRules[1].selectorText')
FAIL document.getElementById('style-container').sheet.cssRules[1].cssText should be :matches(#notthere >> foobar, #target) { color: lime; }. Threw exception TypeError: undefined is not an object (evaluating 'document.getElementById('style-container').sheet.cssRules[1].cssText')
PASS successfullyParsed is true

TEST COMPLETE
